Warn when trying to add folders that contain unsupported filenames
In Qt anything that is not propertly utf8 is transparently not returned,
i.e. QDir will simply pretend the file doesn't exist.

This isn't ideal since people may be asuming a backup is totally done
since no error was reported and one file was not really copied.

This doesn't fix the fact that the file is not copied but at least warns
the file could not by found.

It does so by using KIO::listDir instead of QDir

BUGS: 428115